Refactoring of the configuration
This patch the configuration per buffer of the plugin. Now, with the
disassemble#Disassemble() command, we load the configuration for this
buffer; if not present, we prompt the user to review and validate the
compilation and objdump commands.

This workflow should permit to use different configuration for multiple
files opened at the same time.