Only search the name of the file, without the directory name
When compiling the source file with a relative path, this relative path
is store in the binary file; for `gcc ./test.c -o test -g`, the file
will store `/home/mdedonno/compilation/./test.c` for example as the path
for each source code line.

One solution is this one (only search the file path), the other one
would be to sanitize the output of the objdump command.